# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Enicilion Platform API
  description: Complete OpenAPI specification for Enicilion Enterprise Java Spring Boot Backend.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local Development Server
paths:
...
  /api/events/{slug}:
    get:
      summary: Retrieve an event and its active ticket tiers by slug (Public)
      tags:
        - Events
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
          example: motorscape-2026
      responses:
        '200':
          description: Event detail returned successfully
...
  /api/admin/events:
    get:
      summary: List events for the authenticated admin or staff's organizer (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of events returned successfully
    post:
      summary: Create a new event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
                - location
                - eventDate
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                location:
                  type: string
                eventDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [draft, published, ongoing, completed, cancelled]
                registrationSchema:
                  type: string
                organizerId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Event created successfully
  /api/admin/events/{id}:
    put:
      summary: Update an existing event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
                - location
                - eventDate
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                location:
                  type: string
                eventDate:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [draft, published, ongoing, completed, cancelled]
                registrationSchema:
                  type: string
                organizerId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Event updated successfully
    delete:
      summary: Delete an event and all associated tiers (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Event deleted successfully
  /api/admin/events/{id}/tiers:
    get:
      summary: List all ticket tiers (both public and private) for an event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket tiers list returned successfully
    post:
      summary: Create a ticket tier for an event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
              properties:
                name:
                  type: string
                price:
                  type: number
                quantity:
                  type: integer
                description:
                  type: string
                isPublic:
                  type: boolean
      responses:
        '201':
          description: Ticket tier created successfully
  /api/admin/events/{eventId}/tiers/{tierId}:
    put:
      summary: Update an existing ticket tier (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: tierId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
              properties:
                name:
                  type: string
                price:
                  type: number
                quantity:
                  type: integer
                description:
                  type: string
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Ticket tier updated successfully
    delete:
      summary: Delete a ticket tier (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: eventId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: tierId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket tier deleted successfully
  /api/admin/events/{id}/settle:
    post:
      summary: Perform financial settlement and freeze revenue metrics for a completed event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Financial metrics calculated and frozen successfully
    get:
      summary: Retrieve the frozen settlement summary for an event (Admin / Staff Only)
      tags:
        - Administration
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Settlement summary returned successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...